upd: compatibility

+++ src/picasso.c
@@ -6,7 +6,7 @@
 /*   By: hbalasan <hbalasan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/21 20:35:42 by hbalasan          #+#    #+#             */
-/*   Updated: 2023/09/25 23:27:27 by hbalasan         ###   ########.fr       */
+/*   Updated: 2023/09/26 17:25:12 by hbalasan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,9 +26,12 @@ uint64_t interpolate(int color1, int color2, float fraction)
 	rgb.b1 = color1 & 0xff;
 	rgb.b2 = color2 & 0xff;
 
-	return (int) ((rgb.r2 - rgb.r1) * fraction + rgb.r1) << 16 |
-		(int) ((rgb.g2 - rgb.g1) * fraction + rgb.g1) << 8 |
-		(int) ((rgb.b2 - rgb.b1) * fraction + rgb.b1);
+	// return (int) ((rgb.r2 - rgb.r1) * fraction + rgb.r1) << 16 |
+	// 	(int) ((rgb.g2 - rgb.g1) * fraction + rgb.g1) << 8 |
+	// 	(int) ((rgb.b2 - rgb.b1) * fraction + rgb.b1);
+	return (uint64_t) ((rgb.r2 - rgb.r1) * fraction + rgb.r1) << 16 |
+		(uint64_t) ((rgb.g2 - rgb.g1) * fraction + rgb.g1) << 8 |
+		(uint64_t) ((rgb.b2 - rgb.b1) * fraction + rgb.b1);
 }
 
 void	picasso(t_fdf *fdf, t_coordinates *xyz)
